Fixed some linting erros
Improved database connection in the migration command
