+++ babel.config.js
@@ -18,6 +18,7 @@ module.exports = (api) => {
 				{
 					root: ['./'],
 					alias: {
+						cli: './cli',
 						src: './src',
 						base: './src/base',
 						bootstrap: './src/bootstrap',
