+++ src/base/extensions/cli.d.ts
@@ -1,0 +1,1 @@
+declare module 'cli/commands' {}
