+++ src/base/extensions/database.d.ts
@@ -1,1 +1,4 @@
 declare module 'base/database/drivers' {}
+
+declare module 'database/migrations' {}
+declare module 'database/seeders' {}
