+++ tsconfig.json
@@ -9,6 +9,7 @@
 		"experimentalDecorators": true,
 		"baseUrl": ".",
 		"paths": {
+			"cli/*": ["./cli/*"],
 			"src/*": ["./src/*"],
 			"base/*": ["./src/base/*"],
 			"bootstrap/*": ["./src/bootstrap/*"],
